+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/ServerUpdatePreProcessing.java
@@ -16,7 +16,7 @@ import java.util.List;
 
 public class ServerUpdatePreProcessing {
   private final static Logger LOG = LoggerFactory.getLogger(ServerUpdatePreProcessing.class);
-  private final static List<String> resources = Arrays.asList("PinVol.exe", "vpxtool.exe", "maintenance.jpg", "sendKeys.bat", ScoringDB.SCORING_DB_NAME);
+  private final static List<String> resources = Arrays.asList("PinVol.exe", "nircmd.exe", "vpxtool.exe", "maintenance.jpg", "sendKeys.bat", ScoringDB.SCORING_DB_NAME);
 
   public static void execute() {
     new Thread(() -> {