+++ RELEASE_NOTES.md
@@ -21,7 +21,7 @@
 - Added **vpxtool** from https://github.com/francisdb/vpxtool. This tool is used to import and export VB script from VPX files.
 - **VPX Script Editor(!)**: Finally, thanks to the **vpxtool** you now can not only view the VB script of a table, but also edit it and save it back to the VPX file. The editor for this does not offer much, but it will allow you to do some quick table adjustments.
   
-<img src="https://raw.githubusercontent.com/syd711/vpin-studio/main/documentation/tables/uploads.png" width="600" />
+<img src="https://raw.githubusercontent.com/syd711/vpin-studio/main/documentation/tables/script-viewer.png" width="600" />
 
 - **Server Settings**: Added option to restart the VPin Studio Server.
 - **VPin Studio Installer**: Added registry option to auto-start the server. The installation will no longer ask for the auto-start folder but will use the registry instead. For existing users: you can keep the existing auto-start as is.
@@ -35,3 +35,4 @@
 - **Table Cloning**: Added missing cloning of .ini files.
 - **VPX Commands**: Fixed VPX operations (start/stop/launch to edit) for Visual Pinball installations having a whitespace in the directory path.
 - **Overlay Key Events**: Added button debounce that fixes possible duplicated key events sent by input controllers.
+- **GPU Usage**: Fixed high GPU usage caused by the disabled(!) progress indicator of the jobs dropdown menu (WTF???).